requirements.txt
# ============================================
# Low Engagement Detection System - Requirements
# ============================================
# Python 3.12+
# Install: pip install -r requirements.txt

# ============================================
# Core ML & Data Science
# ============================================
numpy>=1.26.0
pandas>=2.1.0
scikit-learn>=1.3.0
scipy>=1.11.0

# ============================================
# Deep Learning (TensorFlow/Keras)
# ============================================
tensorflow>=2.15.0
keras>=3.0.0
h5py>=3.10.0

# ============================================
# Signal Processing & Change Point Detection
# ============================================
densratio>=0.3.0

# ============================================
# Data Balancing (SMOTE)
# ============================================
imbalanced-learn>=0.11.0

# ============================================
# Web Framework (Backend API)
# ============================================
Flask>=3.0.0
flask-cors>=4.0.0

# ============================================
# Database (MongoDB)
# ============================================
pymongo>=4.6.0

# ============================================
# Sensor Communication (USB Serial)
# ============================================
pyserial>=3.5

# ============================================
# Scheduling & File Watching
# ============================================
schedule>=1.2.0
watchdog>=3.0.0

# ============================================
# Data Visualization
# ============================================
matplotlib>=3.8.0

# ============================================
# Utilities
# ============================================
natsort>=8.4.0
pytz>=2023.3
requests>=2.31.0
python-dotenv>=1.0.0
tabulate>=0.9.0
joblib>=1.3.0

# ============================================
# FastAPI Backend (for annotations)
# ============================================
fastapi>=0.109.0
uvicorn>=0.27.0
python-multipart>=0.0.6
pydantic>=2.5.0

# ============================================
# Development (Optional)
# ============================================
# pytest>=7.0.0
# black>=23.0.0
# flake8>=6.0.0

